Update domain names for RFC 2606 & 2761 compliance
This changes all the domain names from "test.net" to "simp.test."
RFC 2606 designated the "test" TLD as reserved for testing purposes such
as this, and RFC 2761 lays out standards to that-- if followed-- ensure
it remains isolated from the global DNS.  This benefits the environment
by improving resolution time against the global DNS (it will return a
null response right away instead of recursivly resolving to the root
servers), and by eliminating the posibility of resolution to a
third-party address.

Additionally, this changes the host names of the machines from "puppet"
for the server and "client" for the client so "server01" and "client01",
respectively.  The server name more accurately reflects the role of the
machine, and the numbers make it easier to expand the environment with a
uniform naming scheme.