Remove extraneous `higher.domain` DNS zone
This removes configuration for the `higher.domain` zone in BIND DNS, an
empty forwarding zone.  The `.` (wildcard) zone covers it and uses
essentially the same configuration.  The only differnce was that
`higher.domain` did not specify `forward-only`, whereas `.` does.
However, without local zone db files, the distinction was meaningless.
Further, `higher.domain` is not a real domain and its use goes against
RFC 2606.