Make client secondary interfaces static
This adds a profile class in the `site` module for the clients
hostgroup, using it to configure their secondary (private network)
interfaces.  Prior to this the interfaces were un-managed by Puppet,
leaving their configuration up to Vagrant.  Normally that would be best,
allowing hosts to configure their environment as needed, however in this
case we assign addresses using the SIMP DHCP/PXE-boot/Kickstart server
and prefer to manage addresses internally.