(SIMP-1517) Fix client Puppet and Hiera code
Prior to this patch the clients YAML file used an invalid document
seperator and the clients profile did not properly escape its vairables,
causing client catalogs to fail compilation.  This corrects those
problems.

Additionally, it sets the netmask in the clients profile to the netmask
fact for the interface.  It is expected that the DHCP server will set
the IP address and netmask, so the profile just converts those values
into a static config file.

Change-Id: Id33640e80a13722927bff382dbcba1bc8a38bc9e