(SIMP-1517) Configure the server's private interface
Prior to this change the private (second) interface on the server was
not defined in Puppet code.  Packer does not use the second interface,
hence it was not configured during the Packer build.  Normally this
would not be a problem, but some of the Puppet code depends on the
existence of that interface's IP address.  This change adds the
appropriate configuration to ensure the interface is configured and
running if it exists.

Change-Id: Id20614c28a745bd6d6132d4073346e396400e09b